+++ libraries/effect/src/main/java/androidx/media3/effect/InputSwitcher.java
@@ -32,6 +32,7 @@ import androidx.media3.common.ColorInfo;
 import androidx.media3.common.FrameInfo;
 import androidx.media3.common.GlObjectsProvider;
 import androidx.media3.common.GlTextureInfo;
+import androidx.media3.common.OnInputFrameProcessedListener;
 import androidx.media3.common.VideoFrameProcessingException;
 import androidx.media3.common.VideoFrameProcessor;
 import com.google.common.collect.ImmutableList;
@@ -190,20 +191,26 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     checkNotNull(activeTextureManager).setInputFrameInfo(inputFrameInfo);
   }
 
+  /** Returns whether the {@code InputSwitcher} is connected to an active input. */
+  public boolean hasActiveInput() {
+    return activeTextureManager != null;
+  }
+
   /**
    * Returns the {@link TextureManager} that is currently being used.
    *
-   * <p>Must call {@link #switchToInput} before calling this method.
+   * @throws IllegalStateException If the {@code InputSwitcher} is not connected to an {@linkplain
+   *     #hasActiveInput() input}.
    */
   public TextureManager activeTextureManager() {
-    return checkNotNull(activeTextureManager);
+    return checkStateNotNull(activeTextureManager);
   }
 
   /**
    * Invokes {@link TextureManager#signalEndOfCurrentInputStream} on the active {@link
    * TextureManager}.
    */
-  public void signalEndOfCurrentInputStream() {
+  public void signalEndOfInputStream() {
     checkNotNull(activeTextureManager).signalEndOfCurrentInputStream();
   }
 
@@ -218,6 +225,28 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   public Surface getInputSurface() {
     checkState(containsKey(inputs, INPUT_TYPE_SURFACE));
     return inputs.get(INPUT_TYPE_SURFACE).textureManager.getInputSurface();
+  }
+
+  /**
+   * See {@link DefaultVideoFrameProcessor#setInputDefaultBufferSize}.
+   *
+   * @throws IllegalStateException If {@link VideoFrameProcessor#INPUT_TYPE_SURFACE} is not
+   *     {@linkplain #registerInput registered}.
+   */
+  public void setInputDefaultBufferSize(int width, int height) {
+    checkState(containsKey(inputs, INPUT_TYPE_SURFACE));
+    inputs.get(INPUT_TYPE_SURFACE).textureManager.setDefaultBufferSize(width, height);
+  }
+
+  /**
+   * Sets the {@link OnInputFrameProcessedListener}.
+   *
+   * @throws IllegalStateException If {@link VideoFrameProcessor#INPUT_TYPE_TEXTURE_ID} is not
+   *     {@linkplain #registerInput registered}.
+   */
+  public void setOnInputFrameProcessedListener(OnInputFrameProcessedListener listener) {
+    checkState(containsKey(inputs, INPUT_TYPE_TEXTURE_ID));
+    inputs.get(INPUT_TYPE_TEXTURE_ID).textureManager.setOnInputFrameProcessedListener(listener);
   }
 
   /** Releases the resources. */