+++ libraries/effect/src/main/java/androidx/media3/effect/VideoFrameProcessingTaskExecutor.java
@@ -25,7 +25,6 @@ import androidx.media3.common.util.UnstableApi;
 import java.util.ArrayDeque;
 import java.util.Queue;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 import java.util.concurrent.RejectedExecutionException;
@@ -106,25 +105,6 @@ import java.util.concurrent.RejectedExecutionException;
 
     if (executionException != null) {
       handleException(executionException);
-    }
-  }
-
-  /** Submits the given {@link Task} to execute, and returns after the task is executed. */
-  public void submitAndBlock(Task task) {
-    synchronized (lock) {
-      if (shouldCancelTasks) {
-        return;
-      }
-    }
-
-    Future<?> future = wrapTaskAndSubmitToExecutorService(task, /* isFlushOrReleaseTask= */ false);
-    try {
-      future.get();
-    } catch (ExecutionException e) {
-      handleException(e);
-    } catch (InterruptedException e) {
-      Thread.currentThread().interrupt();
-      handleException(e);
     }
   }